[#565] Updated basic_component_py to pyqt5 Added python files to gitignore
