+++ src/tools/basic_component_py/gui/communicator.py
@@ -19,6 +19,6 @@
 #!/usr/bin/python
 # -*- coding: utf-8 -*-
 
-from PyQt4 import QtCore
+from PyQt5 import QtCore
 class Communicator(QtCore.QObject):
     updateBW = QtCore.pyqtSignal()
