+++ src/tools/basic_component_py/resources/resources.qrc
@@ -1,7 +1,5 @@
 <RCC>
   <qresource prefix="images">
     <file>ball.png</file>
-    <file>play.png</file>
-    <file>stop.png</file>
   </qresource>
 </RCC>
